requirements.txt
boto3
numpy
pygame
python-dotenv
sounddevice
RPi.GPIO
Flask
requests